+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Support for `login_hint` query parameter to `GET /login` endpoint.
 - Support for `identifier` query parameter to `GET /login/identifier` endpoint.
 - Support for `loginHint` option to `login` prompt.